Initial commit:
Parsing audio data to find pika calls.

find_peaks is a method found on the internet to duplicate the matlab findpeaks
functionality - see file for attribution/licensing/etc.

pika.py is a port of Shankar's matlab code.
The code may need a little adjustment as I have translated it.  It does seem
to be processing things - however the output file after running on an audio
recording containing pika calls was basically the full file, so perhaps
I did something wrong translating the filtering or something.