Continued work on integrating db models, systematically processing input audio, and identifying a wider variety of pika calls.
As for the wider variety of calls, I found that the previous stricture
on the fundamental harmonic that it be greater than 30 (in the range that
the pre_process limits the window to) was filtering out many calls in some
of the newer files I have been working with so I removed that (but am retaining
the requirement that it be less than 60).  I also have somewhat expanded the
allowed intervals for interpeak distances.